% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_match.R
\name{ext_match}
\alias{ext_match}
\title{Within-folder extension match}
\usage{
ext_match(path_tbl, exts, ...)
}
\arguments{
\item{path_tbl}{Tibble of paths created with make_path_tibble()}

\item{exts}{Atomic character vector of file extensions (excluding period, e.g., "csv") that a folder should contain}

\item{...}{Arguments passed to dplyr::group_by()}
}
\value{
Tibble structure
}
\description{
Within-folder extension match
}
\examples{
ext_match(example_paths, ext = c("csv", "log", "psydat"), main_dir, last_dir)
}
