% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_ids.R
\name{get_df_ids}
\alias{get_df_ids}
\title{Get ID from file}
\usage{
get_df_ids(path_tbl, ext, id_col, read_fun, ...)
}
\arguments{
\item{path_tbl}{Tibble of paths created by make_path_tibble()}

\item{ext}{Atomic character vector of length one identifying file extension to filter by}

\item{id_col}{Atomic character vector of length one identiying ID column in file}

\item{read_fun}{Unquoted function name for reading in files (e.g., read.csv)}

\item{...}{Arguments passed to read function}
}
\value{
A tibble equal in length to the tibble passed in, with a column of IDs added (df_id). If an ID was not found/extracted for the associated file, NA is returned.
}
\description{
Get ID from file
}
\examples{
tmp <- tempdir()
file <- paste0(tmp, .Platform$file.sep, "tmp.csv")
write.csv(data.frame(id = 1), file = file)
tmp_tbl <- make_path_tibble(tmp)
get_df_ids(tmp_tbl, ext = "csv", id_col = "id", read_fun = read.csv)
}
