% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_match.R
\name{id_match}
\alias{id_match}
\title{Compare Folder/File IDs}
\usage{
id_match(x, y, pattern_x, pattern_y)
}
\arguments{
\item{x}{atomic character vector containing ID}

\item{y}{atomic character vector containing ID}

\item{pattern_x}{regular expression used by str_match()}

\item{pattern_y}{regular expression used by str_match()}
}
\value{

}
\description{
Compare Folder/File IDs
}
\examples{
id_match(
example_paths$file,
example_paths$last_dir,
pattern_x = "[0-9]{4,5}",
pattern_y = "[0-9]{4,5}"
)
}
