% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_folder_file_count.R
\name{in_folder_file_count}
\alias{in_folder_file_count}
\title{Count files in folder}
\usage{
in_folder_file_count(path_tbl, ...)
}
\arguments{
\item{path_tbl}{tibble of paths created with make_path_tibble()}

\item{...}{Arguments passed to group_by()}
}
\value{
Tibble structure
}
\description{
Count files in folder
}
\examples{
in_folder_file_count(example_paths, main_dir, last_dir)
}
