% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_folder_match.R
\name{in_folder_match}
\alias{in_folder_match}
\title{Within-Folder pattern match}
\usage{
in_folder_match(path_tbl, pattern, ...)
}
\arguments{
\item{path_tbl}{Tibble of paths created with make_path_tibble()}

\item{pattern}{Regular expression passed to stringr::str_extract()}

\item{...}{Arguments passed to group_by()}
}
\value{
Tibble structure
}
\description{
Within-Folder pattern match
}
\examples{
in_folder_match(example_paths, "^[0-9]{5}", main_dir, last_dir)
}
