% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_all.R
\name{intersect_all}
\alias{intersect_all}
\title{Intersect all vectors}
\usage{
intersect_all(...)
}
\arguments{
\item{...}{atomic numeric vectors}
}
\value{
atomic vector of intersection of vectors
}
\description{
Intersect all vectors
}
\examples{
intersect_all(1:5, 3:6, 2:8)
}
