% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_path_tibble.R
\name{make_path_tibble}
\alias{make_path_tibble}
\title{Tibble of file information}
\usage{
make_path_tibble(paths, patterns = NULL, ...)
}
\arguments{
\item{paths}{Character vector (atomic or list) of top-level file paths for which you'd like to create a tibble structure.}

\item{patterns}{Character vector (atomic or list) of file patterns passed to list.files(). Length must either be one or equal to length of paths.}

\item{...}{Arguments passed to list.files()}
}
\value{
Tibble structure of file and folder information
}
\description{
Tibble of file information
}
\examples{
make_path_tibble(getwd(), recursive = FALSE)
}
