% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pattern.R
\name{match_pattern}
\alias{match_pattern}
\title{Pattern matching}
\usage{
match_pattern(x, pattern)
}
\arguments{
\item{x}{Atomic character vector}

\item{pattern}{regular expression passed to stringr::str_detect()}
}
\value{
Either a message if all elements match, or a character vector of mismatches
}
\description{
Pattern matching
}
\examples{
match_pattern(example_paths$last_dir, "ID_[0-9]{5}")
match_pattern(example_paths$file, "^[0-9]{5}")
}
