% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersect_all.R
\name{outersect_all}
\alias{outersect_all}
\title{Outersect of all vectors}
\usage{
outersect_all(x, names = NULL)
}
\arguments{
\item{x}{A list, either named or unnamed, of numeric vectors}

\item{names}{Optional character vector of names for x}
}
\value{
List of numeric vectors containing the outersect
}
\description{
Outersect of all vectors
}
\examples{
outersect_all(list(1:5, 3:6, 2:8), names = c("a", "b", "c"))
}
